Add hostgroup creation shell script to repo
The next port will be the hostgroup creation script so that I can add and remove both host and service groups.
Creating the list of hosts still requires manual effort on boxes where sed can't be used to do the job in a few seconds.
I may port the sed one-liners over but it makes far more sense to do this graphically with a checkbox or using sed, then with pure python.