add cmdline args for warning, critical, unknown
In trying to keep things as modular as possible I have implemented commandline switches for warning, critical, and unknown.  By default they are set to their respective vaules and nothing will change, but when used you can control the alerts displayed on the dashboard as welll as the notifications a little eaiser.  Certain environments may only want to see warnings or critical, in some cases you can set the value the same but this is not always true, and races conditions may also happen, so to avoid ambiguity, I added these switches to this service check and going forward they will be a standrad part of all service checks, or any other script that reports state information to a monitoring server.