Make the warn, crit, unknown conditionals explict
I changed the code in the warning, critical, and unknown arguements to set the variables explictly instead of implicitly.  It makes for more code but where this is a crucial conditional, I don't want any ambiguity in the code.   It should only do x when y is present, not just if !z is present.