Port check_load to Python
Port Nagios Plugins check_load to Python.  The plugin is nearly functionally identical to the Nagios Plugin version.  The differences are that is adds the ability to select if the user wants to enable warnings, critical, or unknows.  It also displays the time the check to to execute in the perf data.

The nagios Plugin would read the information from uptime, I decided to go directly to /proc/loadavg as I prefer to use first-hand data whenever possible.